location_analysis/requirements.txt
python-polylabel
dash_core_components
dash_html_components
numpy
plotly
scikit-learn